requirements.txt
flask
gunicorn
python-dotenv
Flask-SQLAlchemy
SQLAlchemy
flask-cors
instagrapi
pillow
flask-restx
autopep8
pylint